problem 10 is now working. added tuning definition to gen primes common utils

+++ common_utls.h
@@ -5,6 +5,6 @@ int isPrime(int num);
 
 int isPalindrome(int num);
 
-char*  genPrime(int index);
+char*  genPrime(long index);
 
 #endif
